feat (api): scheduler cleaner keep a number of executions

+++ engine/api/scheduler/cleaner.go
@@ -10,15 +10,16 @@ import (
 )
 
 //Cleaner is the cleaner main goroutine
-func Cleaner(delay time.Duration) {
+func Cleaner(nbToKeep int) {
 	for {
+		CleanerRun(nbToKeep)
 		time.Sleep(10 * time.Minute)
-		CleanerRun(delay)
 	}
 }
 
 //CleanerRun is the core function of the cleaner goroutine
-func CleanerRun(delay time.Duration) ([]sdk.PipelineSchedulerExecution, error) {
+func CleanerRun(nbToKeep int) ([]sdk.PipelineSchedulerExecution, error) {
+	log.Debug("CleanerRun> Deleting old executions...")
 	_db := database.DB()
 	if _db == nil {
 		return nil, fmt.Errorf("Database is unavailable")
@@ -33,20 +34,36 @@ func CleanerRun(delay time.Duration) ([]sdk.PipelineSchedulerExecution, error) {
 
 	//Starting with exclusive lock on the table
 	if err := LockPipelineExecutions(tx); err != nil {
+		log.Debug("CleanerRun> Unable to take lock : %s", err)
 		return nil, err
 	}
 
-	log.Debug("CleanerRun> Deleting old executions...")
-
-	t := time.Now().Add(delay * -1)
-	exs, err := LoadPastExecutions(tx, t)
+	//Load schedulers
+	ps, err := LoadAll(tx)
 	if err != nil {
+		log.Warning("CleanerRun> Unable to load pipeline schedulers : %s", err)
 		return nil, err
 	}
 
-	for i := range exs {
-		if err := DeleteExecution(tx, &exs[i]); err != nil {
-			log.Critical("CleanerRun> Unable to delete execution %d", exs[i].ID)
+	deleted := []sdk.PipelineSchedulerExecution{}
+	for _, s := range ps {
+		exs, err := LoadPastExecutions(tx, s.ID)
+		if err != nil {
+			log.Warning("CleanerRun> Unable to load pipeline schedulers execution : %s", err)
+			return nil, err
+		}
+
+		nbToDelete := len(exs) - nbToKeep
+		nbDeleted := 0
+		for i := range exs {
+			if nbDeleted > nbToDelete {
+				break
+			}
+			if err := DeleteExecution(tx, &exs[i]); err != nil {
+				log.Critical("CleanerRun> Unable to delete execution %d", exs[i].ID)
+			}
+			nbDeleted++
+			deleted = append(deleted, exs[i])
 		}
 	}
 
@@ -55,5 +72,5 @@ func CleanerRun(delay time.Duration) ([]sdk.PipelineSchedulerExecution, error) {
 		return nil, err
 	}
 
-	return exs, nil
+	return deleted, nil
 }
